Move StringGenerator to internal pkg
