+++ pkg/provider/gardener/ruleset/disak8sstig/v1r10/242399_test.go
@@ -26,6 +26,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	fakeclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
+	fakestrgen "github.com/gardener/diki/pkg/internal/stringgen/fake"
 	"github.com/gardener/diki/pkg/kubernetes/pod"
 	fakepod "github.com/gardener/diki/pkg/kubernetes/pod/fake"
 	"github.com/gardener/diki/pkg/provider/gardener/ruleset/disak8sstig/v1r10"
@@ -60,7 +61,7 @@ var _ = Describe("#242399", func() {
 	)
 
 	BeforeEach(func() {
-		v1r10.Generator = &FakeRandString{CurrentChar: 'a'}
+		v1r10.Generator = &fakestrgen.FakeRandString{Rune: 'a'}
 		fakeControlPlaneClient = fakeclient.NewClientBuilder().WithScheme(kubernetesgardener.SeedScheme).Build()
 		fakeClusterClient = fakeclient.NewClientBuilder().WithScheme(kubernetesgardener.ShootScheme).Build()
