+++ pkg/provider/gardener/ruleset/disak8sstig/v1r10/node_files_permissions_and_owner_test.go
@@ -18,6 +18,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	fakeclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
 
+	fakestrgen "github.com/gardener/diki/pkg/internal/stringgen/fake"
 	"github.com/gardener/diki/pkg/kubernetes/pod"
 	fakepod "github.com/gardener/diki/pkg/kubernetes/pod/fake"
 	"github.com/gardener/diki/pkg/provider/gardener/ruleset/disak8sstig/v1r10"
@@ -71,7 +72,7 @@ var _ = Describe("#RuleNodeFiles", func() {
 	)
 
 	BeforeEach(func() {
-		v1r10.Generator = &FakeRandString{CurrentChar: 'a'}
+		v1r10.Generator = &fakestrgen.FakeRandString{Rune: 'a'}
 		fakeClusterClient = fakeclient.NewClientBuilder().Build()
 		fakeControlPlaneClient = fakeclient.NewClientBuilder().WithScheme(kubernetesgardener.SeedScheme).Build()
