+++ pkg/provider/gardener/ruleset/disak8sstig/v1r10/stringgenerator.go
@@ -5,39 +5,10 @@
 package v1r10
 
 import (
-	"math/rand"
-	"time"
+	"github.com/gardener/diki/pkg/internal/stringgen"
 )
-
-type StringGenerator interface {
-	Generate(int) string
-}
 
 var (
-	chars = []rune("1234567890abcdefghijklmnopqrstuvwxyz")
 	// Generator is a not secure random Generator. Exposed for testing purposes.
-	Generator = StringGenerator(NewRand(rand.NewSource(time.Now().UnixNano()), chars))
+	Generator stringgen.StringGenerator = stringgen.Default()
 )
-
-// Generate generates a not secure random string with length n.
-func (r *RandString) Generate(n int) string {
-	runes := make([]rune, n)
-	for i := range runes {
-		runes[i] = r.chars[r.random.Intn(len(r.chars))]
-	}
-	return string(runes)
-}
-
-// RandString is not a secure random generator.
-type RandString struct {
-	random *rand.Rand
-	chars  []rune
-}
-
-// NewRand returns a not secure random string generator.
-func NewRand(source rand.Source, chars []rune) *RandString {
-	return &RandString{
-		random: rand.New(source), //nolint:gosec
-		chars:  chars,
-	}
-}
