+++ pkg/provider/gardener/ruleset/disak8sstig/v1r11/v1r11_suite_test.go
@@ -22,16 +22,3 @@ func TestV1R11(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "DISA Kubernetes STIG V1R11 Test Suite")
 }
-
-func (r *FakeRandString) Generate(n int) string {
-	b := make([]rune, n)
-	for i := 0; i < n; i++ {
-		b[i] = r.CurrentChar
-	}
-	r.CurrentChar++
-	return string(b)
-}
-
-type FakeRandString struct {
-	CurrentChar rune
-}
